--- api/src/controller/payment/get.Success.ts
@@ -1,5 +1,0 @@
-const getSuccess = () => {
-    return "jelow"
-}
-
-export default getSuccess
